Enforce a search order for settings files
Prior to Python 3.7 dictionaries are unordered.  This leads to unpredictable results when you define a settings.ini file and you already have a .env file for other purposes (such as pipenv environment variables).  Sometimes the settings.ini file will be loaded by decouple and sometimes the .env will be loaded.  This change replaces the Python dict with an OrderedDict, which is compatible back to Python 2.7.